@@ -1,13 +1,10 @@
-import 'package:meta/meta.dart';
-
 class BraintreeDropInResult {
   const BraintreeDropInResult({
-    @required this.paymentMethodNonce,
-    @required this.deviceData,
+    required this.paymentMethodNonce,
+    required this.deviceData,
   });
 
   factory BraintreeDropInResult.fromJson(dynamic source) {
-    if (source == null) return null;
     return BraintreeDropInResult(
       paymentMethodNonce:
           BraintreePaymentMethodNonce.fromJson(source['paymentMethodNonce']),
@@ -19,19 +16,18 @@ class BraintreeDropInResult {
   final BraintreePaymentMethodNonce paymentMethodNonce;
 
   /// String of device data. `null`, if `collectDeviceData` was set to false.
-  final String deviceData;
+  final String? deviceData;
 }
 
 class BraintreePaymentMethodNonce {
   const BraintreePaymentMethodNonce({
-    @required this.nonce,
-    @required this.typeLabel,
-    @required this.description,
-    @required this.isDefault,
+    required this.nonce,
+    required this.typeLabel,
+    required this.description,
+    required this.isDefault,
   });
 
   factory BraintreePaymentMethodNonce.fromJson(dynamic source) {
-    if (source == null) return null;
     return BraintreePaymentMethodNonce(
       nonce: source['nonce'],
       typeLabel: source['typeLabel'],
